requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.12.4
geojson-rewind==1.0.1
idna==3.2
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.2
pyproj==3.2.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
tweepy==3.10.0
urllib3==1.26.6
